cbusアドレスを20bitまで使うように変更

+++ c86usb.cydsn/cbus.h
@@ -3,8 +3,8 @@
 #define CBUS_H__
 
 void cbus_reset(void);
-void cbus_write(uint8_t slot, uint16_t addr, uint16_t data);
-uint16_t cbus_read(uint8_t slot, uint16_t addr);
+void cbus_write(uint8_t slot, uint32_t addr, uint16_t data);
+uint16_t cbus_read(uint8_t slot, uint32_t addr);
 
 
 #endif
